Make phone number mandatory for debt collection
For debt collection complaints, consumers need to enter both the phone
number the collector has been calling AND one of the other identifiers
(account number, loan number, last 4 digits of SSN) in step 4.

To make that happen, we need to pull the phone number out of the list of
identifiers and up to its own section.