Restructure gh-pages for releases
Now that we're doing all our development on the master branch, gh-pages
is going to become a "production" branch where we release the latest
version at the end of each sprint.

This commit restructures the gh-pages branch to get it ready to be a
"production" branch and adds the latest release, 0.52.