Add recreated current landing page for testing
We need to recreate the current submit a complaint landing page
(https://www.consumerfinance.gov/complaint/) for an upcoming user test.

This recreated page will be used in a test of the current complaint
submission experience instead of the live page to (a) allow us to run
the test after the current live page is replaced with the new version
and (b) to allow users to submit complaints to our development
environment.

This commit adds the recreated version of the current submit a complaint
landing page at at
https://cfpb.github.io/complaint-intake/baseline-test/.